# docs/api/api_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Template HTTP API documentation.",
        "title": "Go Template API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/app/user/logout": {
            "post": {
                "description": "Clear the app access token cookie.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_http.Response"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "go_template_access_token=; Path=/api/v1/app; Max-Age=0; HttpOnly; SameSite=Lax"
                            }
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "description": "Create a new user account and issue an app access token cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "Create user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_chi.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-template_internal_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-template_domain_user.User"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "go_template_access_token=...; Path=/api/v1/app; HttpOnly; SameSite=Lax"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_http.Response"
                        }
                    },
                    "409": {
                        "description": "user already exists",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_http.Response"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_http.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-template_domain_user.User": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password_hash": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "go-template_internal_http.Message": {
            "type": "string",
            "enum": [
                "success",
                "fail",
                "too many request",
                "not found",
                "conflict",
                "unauthorized",
                "invalid input",
                "forbidden",
                "service unavailable",
                "Database Error",
                "Internal Server Error"
            ],
            "x-enum-varnames": [
                "Success",
                "Fail",
                "TooManyReq",
                "NotFound",
                "Conflict",
                "UnAuthorized",
                "InvalidInput",
                "Forbidden",
                "ServiceUnavailable",
                "DatabaseError",
                "InternalError"
            ]
        },
        "go-template_internal_http.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/go-template_internal_http.Message"
                }
            }
        },
        "internal_http_chi.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        }
    }
}